shared/dagster_scripts/pyproject.toml
[tool.poetry]
name = "dagster_scripts"
version = "0.0.0"
description = "Scripts for running one-time production jobs."
authors = ["Jasper Ginn"]
readme = "README.md"
packages = [{include = "dagster_scripts", from="src"}]

[tool.poetry.scripts]
dagster_scripts = "dagster_scripts.cli:entrypoint"

[tool.poetry.dependencies]
python = "^3.9,<3.12"
typer = "^0.9.0"
pydantic = "^2.6.1"
dagster = "^1.6.5"
pendulum = "<3.0.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.26.0"

[tool.poetry-git-version-plugin]
alpha_version_format = '{version}a{distance}+{commit_hash}'
version_format = '{version}'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
